Support autocomplete from a group of layers
The idea here is to have several layers which share a set of tags. This
mirrors what we already have if several layers reference the same type
of entity. You might have config like this:

        {
            "id": "keywords",
            "title": "Keywords",
            "canAddAnnotations": {"member": true, "staff": true, "admin": true},
            "item": "Keyword",
            "overlap": true,
            "type": "string",
            "autocomplete": true,
            "autocompleteKeys": ["keywords", "minorkeywords"]
        },
        {
            "id": "minorkeywords",
            "title": "Minor Keywords",
            "canAddAnnotations": {"member": true, "staff": true, "admin": true},
            "item": "Keyword",
            "overlap": true,
            "type": "string",
            "autocomplete": true,
            "autocompleteKeys": ["keywords", "minorkeywords"]
        },

Now, adding new keywords in either bin will offer autocompletions from
the union of the two layers. The other option would be to do this on the
server side, but I thought this was a less invasive way to achieve this.